feat: updated ordner structure for improved overview
