+++ xact_frontend/lib/screens/playnow_screen.dart
@@ -1,0 +1,187 @@
+import 'package:flutter/material.dart';
+
+class PlayNowScreen extends StatelessWidget {
+  const PlayNowScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color(0xFF1A1F2E),
+      body: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 24.0),
+          child: Column(
+            children: [
+              const SizedBox(height: 40),
+              _buildLogo(),
+              const SizedBox(height: 16),
+              const Text(
+                'X-ACT',
+                style: TextStyle(
+                  fontSize: 48,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.white,
+                  letterSpacing: 2,
+                ),
+              ),
+              const SizedBox(height: 8),
+              const Text(
+                'Digital Scotland Yard',
+                style: TextStyle(
+                  fontSize: 18,
+                  color: Colors.white70,
+                  letterSpacing: 1,
+                ),
+              ),
+              const SizedBox(height: 12),
+              const Text(
+                'Hunt down Mister X in this real-world chase game',
+                style: TextStyle(fontSize: 14, color: Colors.white54),
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 32),
+              // Start a New Game Card
+              _buildStartNewGameCard(context),
+              const SizedBox(height: 16),
+              // Join a Friend's Game Card
+              _buildJoinGameCard(context),
+              const Spacer(),
+              // Back Button
+              _buildBackButton(context),
+              const SizedBox(height: 24),
+              const Text(
+                'Play responsibly • Stay safe • Follow local laws',
+                style: TextStyle(fontSize: 12, color: Colors.white38),
+              ),
+              const SizedBox(height: 16),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildLogo() {
+    return Container(
+      width: 80,
+      height: 80,
+      decoration: BoxDecoration(
+        shape: BoxShape.circle,
+        border: Border.all(color: const Color(0xFFE53935), width: 3),
+      ),
+      child: const Center(
+        child: Icon(Icons.location_on, color: Color(0xFFE53935), size: 40),
+      ),
+    );
+  }
+
+  Widget _buildStartNewGameCard(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      decoration: BoxDecoration(
+        color: const Color(0xFF3D5AFE),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Material(
+        color: Colors.transparent,
+        child: InkWell(
+          onTap: () {
+            // TODO: Navigate to create game screen
+          },
+          borderRadius: BorderRadius.circular(12),
+          child: Padding(
+            padding: const EdgeInsets.all(20),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const Icon(Icons.add, color: Colors.white, size: 28),
+                const SizedBox(width: 16),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: const [
+                      Text(
+                        'Start a New Game',
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 18,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      SizedBox(height: 8),
+                      Text(
+                        "You'll be the host and get a code to share with your friends so they can join your game room",
+                        style: TextStyle(color: Colors.white70, fontSize: 14),
+                      ),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildJoinGameCard(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      decoration: BoxDecoration(
+        color: const Color(0xFF252A3A),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Material(
+        color: Colors.transparent,
+        child: InkWell(
+          onTap: () {
+            // TODO: Navigate to join game screen
+          },
+          borderRadius: BorderRadius.circular(12),
+          child: Padding(
+            padding: const EdgeInsets.all(20),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const Icon(Icons.arrow_forward, color: Colors.white, size: 28),
+                const SizedBox(width: 16),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: const [
+                      Text(
+                        "Join a Friend's Game",
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 18,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      SizedBox(height: 8),
+                      Text(
+                        'Enter the code your friend shared with you to join their game room',
+                        style: TextStyle(color: Colors.white70, fontSize: 14),
+                      ),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildBackButton(BuildContext context) {
+    return TextButton(
+      onPressed: () {
+        Navigator.of(context).pop();
+      },
+      child: const Text(
+        'Back',
+        style: TextStyle(color: Colors.white70, fontSize: 16),
+      ),
+    );
+  }
+}
