+++ xact_frontend/lib/screens/startscreen/start_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:xact_frontend/screens/startscreen/get_started.dart';
 import 'package:xact_frontend/screens/playnow_screen.dart';
 
-class GetStartedScreen extends StatelessWidget {
-  const GetStartedScreen({super.key});
+class StartScreen extends StatelessWidget {
+  const StartScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +14,7 @@ class GetStartedScreen extends StatelessWidget {
           padding: const EdgeInsets.symmetric(horizontal: 24.0),
           child: Column(
             children: [
-              const SizedBox(height: 24),
+              const Spacer(flex: 2),
               _buildLogo(),
               const SizedBox(height: 16),
               const Text(
@@ -26,7 +27,6 @@ class GetStartedScreen extends StatelessWidget {
                 ),
               ),
               const SizedBox(height: 8),
-              // Subtitle
               const Text(
                 'Digital Scotland Yard',
                 style: TextStyle(
@@ -36,28 +36,21 @@ class GetStartedScreen extends StatelessWidget {
                 ),
               ),
               const SizedBox(height: 12),
-              // Description
               const Text(
                 'Hunt down Mister X in this real-world chase game',
                 style: TextStyle(fontSize: 14, color: Colors.white54),
                 textAlign: TextAlign.center,
               ),
-              const SizedBox(height: 24),
-              // Hide Help Button
-              _buildHideHelpButton(context),
+              const Spacer(flex: 2),
+              _buildHowToGetStartedButton(context),
               const SizedBox(height: 16),
-              // Steps Card
-              _buildStepsCard(),
-              const Spacer(),
-              // Play Now Button
               _buildPlayNowButton(context),
               const SizedBox(height: 24),
-              // Footer text
               const Text(
                 'Play responsibly • Stay safe • Follow local laws',
                 style: TextStyle(fontSize: 12, color: Colors.white38),
               ),
-              const SizedBox(height: 16),
+              const Spacer(),
             ],
           ),
         ),
@@ -79,13 +72,16 @@ class GetStartedScreen extends StatelessWidget {
     );
   }
 
-  Widget _buildHideHelpButton(BuildContext context) {
+  Widget _buildHowToGetStartedButton(BuildContext context) {
     return SizedBox(
       width: double.infinity,
       height: 56,
       child: OutlinedButton.icon(
         onPressed: () {
-          Navigator.of(context).pop();
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => const GetStartedScreen()),
+          );
         },
         style: OutlinedButton.styleFrom(
           side: const BorderSide(color: Colors.white24),
@@ -96,68 +92,10 @@ class GetStartedScreen extends StatelessWidget {
         ),
         icon: const Icon(Icons.help_outline, color: Colors.white70),
         label: const Text(
-          'Hide Help',
+          'How to Get Started',
           style: TextStyle(color: Colors.white, fontSize: 16),
         ),
       ),
-    );
-  }
-
-  Widget _buildStepsCard() {
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.all(20),
-      decoration: BoxDecoration(
-        color: const Color(0xFF252A3A),
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Column(
-        children: [
-          _buildStepItem(1, 'Create a new game or join a friends game'),
-          const SizedBox(height: 16),
-          _buildStepItem(2, 'Choose your team (Mister X or Detectives)'),
-          const SizedBox(height: 16),
-          _buildStepItem(3, 'Wait for all players to pick their team'),
-          const SizedBox(height: 16),
-          _buildStepItem(4, 'Host starts the game and the fun begins!'),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildStepItem(int number, String text) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-          width: 28,
-          height: 28,
-          decoration: const BoxDecoration(
-            color: Color(0xFFE53935),
-            shape: BoxShape.circle,
-          ),
-          child: Center(
-            child: Text(
-              number.toString(),
-              style: const TextStyle(
-                color: Colors.white,
-                fontWeight: FontWeight.bold,
-                fontSize: 14,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(width: 12),
-        Expanded(
-          child: Padding(
-            padding: const EdgeInsets.only(top: 4),
-            child: Text(
-              text,
-              style: const TextStyle(color: Colors.white, fontSize: 15),
-            ),
-          ),
-        ),
-      ],
     );
   }
